Fix compilation issues GCC

+++ source/blender/blenfont/intern/blf_font.cc
@@ -316,7 +316,7 @@ void blf_batch_draw()
   }
 
   blender::gpu::Texture *texture = blf_batch_cache_texture_load();
-  GPU_storagebuf_usage_size_set(g_batch.glyph_buf, g_batch.glyph_len * sizeof(GlyphQuad));
+  GPU_storagebuf_usage_size_set(g_batch.glyph_buf, size_t(g_batch.glyph_len) * sizeof(GlyphQuad));
   GPU_storagebuf_update(g_batch.glyph_buf, g_batch.glyph_data);
   GPU_storagebuf_bind(g_batch.glyph_buf, 0);
